mark card bill as paid, card delete confirmation, initialise after login

+++ src/app/card/card-list/card-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { Subscription } from 'rxjs';
 import { Icard } from '../card';
 import { CardService } from '../card.service';
@@ -13,7 +14,10 @@ export class CardListComponent implements OnInit, OnDestroy {
   cards: Icard[] = []
   sub!: Subscription;
   aggregate: any = {};
-  constructor( private cardService: CardService) { }
+    
+    constructor( private cardService: CardService,
+      private confirmationService: ConfirmationService,
+      private messageService: MessageService) { }
 
   ngOnInit(): void {
     this.initialise();
@@ -34,8 +38,53 @@ export class CardListComponent implements OnInit, OnDestroy {
     })
   }
 
-  async deleteCard(card: Icard) {
-    await this.cardService.deleteCard(card);
+  async deleteCard(event: Event, card: Icard) {
+
+    this.confirmationService.confirm({
+      target: event.target!,
+      message: 'Are you sure that you delete card: ' + card.cardName,
+      icon: 'pi pi-exclamation-triangle',
+      accept: async () => {
+        this.messageService.add({
+          severity: 'success',
+          summary: 'Success',
+          detail: card.cardName + ' marked for deletion.',
+        });
+        await this.cardService.deleteCard(card);
+      },
+      reject: () => {
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Rejected',
+          detail: 'Card deletion Cancelled.',
+        });
+      },
+    });
+  }
+
+  async markAsPaid(event: Event, card: Icard) {
+
+    this.confirmationService.confirm({
+      target: event.target!,
+      message: 'Are you sure to mark bill as paid for: ' + card.cardName,
+      icon: 'pi pi-exclamation-triangle',
+      accept: async () => {
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Info',
+          detail: card.cardName + ' marked for Updation.',
+        });
+        card.amountDue = 0;
+        await this.cardService.updateCard(card);
+      },
+      reject: () => {
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Rejected',
+          detail: 'Card Updation Cancelled.',
+        });
+      },
+    });
   }
 
   ngOnDestroy(): void {
